arch/run: simplify copy-paste of the command line
The major annoyance coming from errata was "-initrd `mktemp`" that had
to be removed before running the copy-pasted command line.
Move "-initrd ..." part at the end of the line and hide it behind "#" if
the initird file is a temporary one.

This patch also brings s390 to use the same code as other architectures
and eliminates code duplication.

Signed-off-by: Radim KrÄmÃ¡Å™ <rkrcmar@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>