Replace -Wextra with a saner list of warning flags
Using -Wextra together with -Werror is troublesome - various versions
of GCC produce suspicious or even wrong warnings with -Wextra which
then become fatal errors with -Werror. For example, the current state
of the kvm-unit-tests does not compile anymore with GCC 4.8.1 for
s390x due to an inadequate -Wmissing-field-initializers warning.
That's annoying for users who just would like to compile the
kvm-unit-tests and cumbersome for the developers who have to work
around these problems in the source code. So let's replace -Wextra
by a saner lists of warning flags that are normally enabled by -Wextra.
Since they apparently can be used for building x86, too, the flags
are now also applied to the global CFLAGS instead of specifying them
for the single targets only.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Message-Id: <1498761773-11164-3-git-send-email-thuth@redhat.com>
[Paolo: dropped -Wtype-limits, since testing >= 0 is common with the
 0 hidden behind a macro.]
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>