arm64: timer: Fix test on APM X-Gene
When running the vtimer test on an APM X-Gene, setting the timer value
to (2^64 - 1) apparently results in the timer always firing, even
thought the counter is mich lower than the cval.

Since the idea of the code is to set everything up and schedule the
timer for some time very far in the future, take a pragmatic approach
and just add 10s worth of delay instead.

Signed-off-by: Christoffer Dall <cdall@linaro.org>
Message-Id: <20170713192009.10069-3-cdall@linaro.org>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>