x86: hyperv_stimer: add test for busy message slot
When a SynIC timer expires it attempts to deliver a message to the
corresponding slot, which may be busy (e.g. because the guest hasn't yet
processed the previous message).

KVM used to livelock here, endlessly retrying message delivery without
letting the guest to run and release the slot (a patch fixing that has
been posted to KVM ml).

Add a testcase for this scenario.

Signed-off-by: Roman Kagan <rkagan@virtuozzo.com>