arm64: timer: Avoid IRQ race in timer test
The current timer test relies on testing the pending state of the timer
before the interrupt handler has run which could lower the pending
signal again (because it masks the timer output signal).

What we really want is to make sure the output signal from the timer as
perceived by the virtual interrupt controller is low when the timer is
programmed some time far in the future.  The proper way to do that is to
disable the timer interrupt on the distributor and then reading its
pending state.

Signed-off-by: Christoffer Dall <cdall@linaro.org>
Message-Id: <20170726114249.17774-1-cdall@linaro.org>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>