x86: access: fix SMEP-disabled case with TCG
TCG always allows setting CR4.SMEP, even when the CPUID bit is
disabled.  In this case, however, the test causes a triple fault
because it tries to execute a user page with CR4.SMEP=1.  The
right way to do the test is to go through set_cr4_smep.  Change
set_cr4_smep to return whether an exception happened, and use it
in ac_test_run.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>