x86: ioapic: Test self-reconfiguration from interrupt handler
Make sure that we can reconfigure the destination of an irq from the
interrupt handler. Make sure that the EOI is correctly delivered to
the IOAPIC and that remote_irr is reset to 0.

This tests the fix introduced by commit "KVM: x86: ioapic: Fix
level-triggered EOI and IOAPIC reconfigure race".

Signed-off-by: Nikita Leshenko <nikita.leshchenko@oracle.com>
Reviewed-by: Liran Alon <liran.alon@oracle.com>
Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Message-Id: <20171107080933.20924-1-nikita.leshchenko@oracle.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>