x86: setup: fixes to alloc_phys setup
There were two thinkos in commit 937e239 ("x86: use alloc_phys",
2017-11-02).

First, alloc_phys is not initialized if there are no multiboot modules,
as is the case when invoking QEMU directly with a test kernel.

Second, one "&" was missing, so the topmost memory address was passed
as the size to phys_alloc_init.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>