arm/arm64: setup: don't allow gaps in phys range
Set the basic range parameters PHYS_OFFSET and PHYS_END to the
primary memory region range parameters to ensure there are no
gaps, as the regions in the device tree may not be contiguous.
If a unit test wants to use memory in another region, then it
can get it from mem_regions.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>