vmx: fixes for MAXPHYADDR requirement
The EPT access tests require at least 41 bits of guest physical address
space.  This is because the test data needs a separate PML4 entry in the
EPT table.  However, many consumer-grade processors have only 36 or
39-bit MAXPHYADDR, and the tests fail there with a page fault due to
reserved bits set in the guest (GVA->GPA) page tables.

Add a test on MAXPHYADDR and skip the tests if they cannot run on the
system under test, and only require 40 bits to lower the requirements.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>