x86: prevent GCC from using sse* instructions
GCC 8 emitted MOVDQU when compiling vmx.elf, but we do not enable
CR4.OSFXSR in that test, so the instruction throws #UD.
This patch forbids sse instructions, instead of enabling CR4.OSFXSR, as
I think it's better to keep the environment minimal.

Paolo suggested a way to keep sse around the function that uses it,
allowing us to disable all sse instructions.

Reviewed-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>