s390x: Add privileged operation exceptions handling
A huge part of the emulated instructions are privileged and result in
an operation exception, when executed in problem state.

To be able to test this, we introduce the enter_pstate function, that
sets the cpu to the problem state. Also we extend the interrupt
handler, so it resets us to supervisor mode again, when we hit the
privileged operation exception.

Signed-off-by: Janosch Frank <frankja@linux.vnet.ibm.com>
Reviewed-by: Thomas Huth <thuth@redhat.com>
Reviewed-by: David Hildenbrand <david@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>