s390x: Add simple guarded storage tests
Guarded storage was introduced with IBM z14. Loads with the special
instruction LOAD GUARDED are suspect to access controls if they fall
into a guarded area. A handler is called if such an access is
detected. The test checks for exceptions if the instructions are
executed without cr2 enablement and tests handler execution on load
guarded.

The initial test case is once again from Martin Schwidefsky with minor
adaptions by me for kvm-unit-tests.

Signed-off-by: Janosch Frank <frankja@linux.vnet.ibm.com>
Reviewed-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>