s390x: Remove .note.gnu.build-id section from the ELF files
While debugging another problem, we noticed that the low memory contains
some bytes which should not be there. Looks like the linker puts the
".note.gnu.build-id" section to address 0 if it has not explicitly
specified in the linker script. ARM and powerpc unit tests are apparently
already using --build-id=none to avoid this problem, so let's simply use
this linker flag in the s390x builds, too.

Reviewed-by: Janosch Frank <frankja@linux.ibm.com>
Reviewed-by: David Hildenbrand <david@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>