scripts: Speedup MAX_SMP check
The current MAX_SMP check launches QEMU for every value of the SMP
parameter to check if the SMP parameter based on the native is higher
than the QEMU maximum supported configuration.

On something like a TX2, which has 224 threads, this takes a very long
time where the test script just sits there with no output.

Reduce the wait time by taking a right shift, reducing the maximum
number of iterations to log2 MAX_SMP.

Signed-off-by: Christoffer Dall <christoffer.dall@arm.com>
Reviewed-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>