arm/arm64: fix gic-ipi test regression
The gic-ipi tests started crashing in a strange way after
commit 9064d89b ("kvm-unit-tests: unlock before calling
return"). Eventually I narrowed it down to the NR_CPUS change
being the problem. I still don't understand what's going on,
but NR_CPUS=511 works and NR_CPUS=512 doesn't. I don't have
time to debug this further right now, so this is just a
temporary hack that appears to fix the regression.

Signed-off-by: Andrew Jones <drjones@redhat.com>