s390x: try !FORCE SCLP read SCP info if FORCED is unknown
Looking at what the kernel does, looks like we should
- Check for more errors
- Try to execute !FORCED read if the FORCED one is unknown

The kernel sets the function code to 0x80 (SCLP_FC_DUMP_INDICATOR). We
don't need that. Also, the control_mask[2] is set to 0x80
(SCLP_CM2_VARIABLE_LENGTH_RESPONSE), something we also don't need as
4k SCCB should be sufficient for the amount of CPUs we expect to be
called with. If this ever changes, we can extend it.

Tested-by: Janosch Frank <frankja@linux.ibm.com>
Reviewed-by: Janosch Frank <frankja@linux.ibm.com>
Signed-off-by: David Hildenbrand <david@redhat.com>
Signed-off-by: Thomas Huth <thuth@redhat.com>