arm/arm64: selftest.vectors-user: clean up PSCI exit
If we're using the PSCI system off unit test exit, as opposed to the
chr-testdev exit, then the vectors-user test will need to be able to
invoke a PSCI call from user mode. That won't work, but we can simply
try again from the exception handler when the user mode attempt fails
in order to complete the exit.

Reviewed-by: Alexandru Elisei <alexandru.elisei@arm.com>
Tested-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>