Re-name 'test_vmx_controls' to 'test_vmx_vmlaunch' and pass the expected error code to it
The error code returned by KVM when vmlaunch fails, is hard-coded in
'test_vmx_controls'. Because that function has been thus far used only for
testing the VMX controls. Re-name the function to 'test_vmx_vmlaunch' and pass
the expected error code to it so that it can be used for testing any VMCS
field.
Add two wrappers around 'test_vmx_vmlaunch' for testing the VMX controls.

Signed-off-by: Krish Sadhukhan <krish.sadhukhan@oracle.com>
Suggested-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>