run_tests: fix command line options handling
When specifying desired tests on the command line, options are ignored
unless specified prior to any desired tests.
E.g.

  $ ./run_tests.sh sthyi -t emulator -j5
  PASS emulator (102 tests)
  PASS sthyi (19 tests)

Note that the output is not in TAP13 format, and the tests are not executed
in parallel.
This patch fixes the issue by switching to getopt instead. Might not be the
optimum solution, but threw in long command line opts as a bonus.

Signed-off-by: Stefan Raspl <raspl@linux.ibm.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>