arm: powerpc: comment halt(code)
We're being a bit tricky when we change the prototype of halt() to
take 'code' as an argument. Doing this gives us a last chance to
see it when inspecting the halted unit test state (it'll be in
r0/x0 for arm/arm64 and r3 for powerpc.) Let's comment this trick
to avoid confusion.

Suggested-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>