x86: APIC: Add test for pending NMIs while NMIs are blocked
Though explicit documentation is difficult to unearth, x86 guarantees
that exactly one NMI will be pended when NMIs are blocked.  The SDM
essentially calls this out in its section on handling NMIs in SMM:

  NMI interrupts are blocked upon entry to the SMI handler. If an NMI
  request occurs during the SMI handler, it is latched and serviced
  after the processor exits SMM. Only one NMI request will be latched
  during the SMI handler.

Add a test to send multiple NMIs from within an NMI handler to verify
that KVM correctly pends exactly *one* NMI when NMIs are blocked.

Cc: Nadav Amit <nadav.amit@gmail.com>
Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>