x86: PMU: Fix PMU counters masking
Intel SDM says that for MSR_IA32_PERFCTR0/1 "the lower-order 32 bits of
each MSR may be written with any value, and the high-order 8 bits are
sign-extended according to the value of bit 31." The current PMU tests
ignored the fact that the high bit is sign-extended.

At the same time, the fixed counters are not limited to 32-bit, but
appear to be limited to the width of the fixed counters (I could not
find clear documentation).

Fix the tests accordingly.

Signed-off-by: Nadav Amit <nadav.amit@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>