gitlab-ci: Run tests with a Fedora docker image
Fedora has a newer version of QEMU - and most notably it has a *working*
version of qemu-system-s390x! So we can finally also run some s390x tests
in the gitlab-ci.

For some unknown reasons, the sieve test is now failing on x86_64,
so I had to disable it. OTOH, the taskswitch2 test now works on
i386, so we can enable this test instead.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>