x86: Support environments without test-devices
Enable to run the tests when test-device is not present (e.g.,
bare-metal). Users can provide the number of CPUs and ram size through
kernel parameters.

On Ubuntu that uses grub, for example, the tests can be run by copying a
test to the boot directory (/boot) and adding a menu-entry to grub
(e.g., by editing /etc/grub.d/40_custom):

  menuentry 'idt_test' {
	set root='[ROOT]'
	multiboot [BOOT_RELATIVE]/[TEST].flat [PARAMETERS]
	module params.initrd
  }

Replace:
  * [ROOT] with `grub-probe --target=bios_hints /boot`
  * [BOOT_RELATIVE] with `grub-mkrelpath /boot`
  * [TEST] with the test executed
  * [PARAMETERS] with the test parameters

params.initrd, which would be located on the boot directory should
describe the machine and tell the test infrastructure that a test
device is not present and boot-loader was used (the bootloader and qemu
deliver test . For example for a 4 core machines with 4GB of
memory:

  NR_CPUS=4
  MEMSIZE=4096
  TEST_DEVICE=0
  BOOTLOADER=1

Since we do not really use E820, using more than 4GB is likely to fail
due to holes.

Finally, do not forget to run update-grub. Remember that the output goes
to the serial port.

Cc: Andrew Jones <drjones@redhat.com>
Signed-off-by: Nadav Amit <nadav.amit@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>