x86: Fix breakage of fw_cfg for 32-bit unit tests
Ensure the fw_cfg overrides are parsed prior consuming any of said
overrides.  fwcfg_get_u() treats zero as a valid overide value, which
is slightly problematic since the overrides are in the .bss and thus
initialized to zero.

Add a limit check when indexing fw_override so that future code doesn't
spontaneously explode.

Cc: Nadav Amit <nadav.amit@gmail.com>
Fixes: 03b1e4570f967 ("x86: Support environments without test-devices")
Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>