travis.yml: Enable running of tests with TCG
Currently the tests at the end of the .travis.yml script are ignored,
since we can not use KVM in the Travis containers. But we can actually
run of some of the kvm-unit-tests with TCG instead, to make sure that
the binaries are not completely broken.
Thus introduce a new TESTS variable that lists the tests which we can
run with TCG. Unfortunately, the ppc64 and s390x QEMUs in Ubuntu also
need some extra love: The ppc64 version only works with the additional
"cap-htm=off" setting. And the s390x package lacks the firmware and
refuses to work unless we provide a fake firmware file here. Any file
works since the firmware is skipped when "-kernel" is used, so we can
simply use one of the pre-existing files in the source tree.

Signed-off-by: Thomas Huth <thuth@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>